requirements.txt
Django==1.11
django-money==0.12.3
olefile==0.44
Pillow==4.3.0
py-moneyed==0.7.0
pytz==2017.3
